Extract common "handlePageChange" to a reusable hook
